# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Who caused the accident.
 */
export type ClaimInputFaultParty = typeof ClaimInputFaultParty[keyof typeof ClaimInputFaultParty];


export const ClaimInputFaultParty = {
  other: 'other',
  self: 'self',
} as const;

export type ClaimInputPreferredTimeSlot = typeof ClaimInputPreferredTimeSlot[keyof typeof ClaimInputPreferredTimeSlot];


export const ClaimInputPreferredTimeSlot = {
  vormittags: 'vormittags',
  nachmittags: 'nachmittags',
  ganztags: 'ganztags',
} as const;

export interface ClaimInput {
  /** Who caused the accident. */
  faultParty: ClaimInputFaultParty;
  /**
     * Damage description.
     * @minLength 1
     */
  description: string;
  /**
     * Object paths of uploaded damage photos.
     * @maxItems 10
     * @items.maxLength 200
     */
  photoPaths: string[];
  /**
     * @minLength 1
     * @maxLength 200
     */
  firstName: string;
  /** @maxLength 320 */
  email: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 4 */
  postalCode: string;
  accidentDate?: string;
  accidentTime?: string;
  accidentLocation?: string;
  opponentInfo?: string;
  preferredDate?: string;
  preferredTimeSlot?: ClaimInputPreferredTimeSlot;
  /** Consent to privacy policy (must be true). */
  consent: true;
}

export interface Claim {
  id: number;
  faultParty: string;
  description: string;
  photoPaths: string[];
  firstName: string;
  email: string;
  phone: string;
  postalCode: string;
  /** @nullable */
  accidentDate?: string | null;
  /** @nullable */
  accidentTime?: string | null;
  /** @nullable */
  accidentLocation?: string | null;
  /** @nullable */
  opponentInfo?: string | null;
  /** @nullable */
  preferredDate?: string | null;
  /** @nullable */
  preferredTimeSlot?: string | null;
  consent: boolean;
  createdAt: string;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}