# lib/api-zod/src/generated/types/claim.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Claim {
  id: number;
  faultParty: string;
  description: string;
  photoPaths: string[];
  firstName: string;
  email: string;
  phone: string;
  postalCode: string;
  /** @nullable */
  accidentDate?: string | null;
  /** @nullable */
  accidentTime?: string | null;
  /** @nullable */
  accidentLocation?: string | null;
  /** @nullable */
  opponentInfo?: string | null;
  /** @nullable */
  preferredDate?: string | null;
  /** @nullable */
  preferredTimeSlot?: string | null;
  consent: boolean;
  createdAt: Date;
}