# lib/api-zod/src/generated/types/claimInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClaimInputFaultParty } from './claimInputFaultParty';
import type { ClaimInputPreferredTimeSlot } from './claimInputPreferredTimeSlot';

export interface ClaimInput {
  /** Who caused the accident. */
  faultParty: ClaimInputFaultParty;
  /**
     * Damage description.
     * @minLength 1
     */
  description: string;
  /**
     * Object paths of uploaded damage photos.
     * @maxItems 10
     * @items.maxLength 200
     */
  photoPaths: string[];
  /**
     * @minLength 1
     * @maxLength 200
     */
  firstName: string;
  /** @maxLength 320 */
  email: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 4 */
  postalCode: string;
  accidentDate?: string;
  accidentTime?: string;
  accidentLocation?: string;
  opponentInfo?: string;
  preferredDate?: string;
  preferredTimeSlot?: ClaimInputPreferredTimeSlot;
  /** Consent to privacy policy (must be true). */
  consent: true;
}